Allow loading iterative manifests into the MoM
When loading a MoM the file gets downloaded then it is parsed.
Currently only bundle manifests with a key of M are being used, so
manifests that use an I (from "Iterative") are being ignored.

This commit enables manifests with an I key to be loaded to memory
when loading the MoM as valid manifests.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>