Enable tests for refactored create_update_list
When creating a MoM that does not have a previous version, set the default
minversion to match the current version. This will enable tests to use
the new create_update_list instead of the old implementation which will
be used as a fallback. Also, this change adds a test to validate the
fallback create_update_list function.

Signed-off-by: John Akre <john.w.akre@intel.com>